% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Dawson}
\alias{Dawson}
\title{compute Dawson(z) = sqrt(pi)/2  *  exp(-z^2) * erfi(z)}
\usage{
Dawson(z_, relerr = 0)
}
\arguments{
\item{relerr}{double, requested error}

\item{z}{complex vector}
}
\value{
complex vector
}
\description{
compute Dawson(z) = sqrt(pi)/2  *  exp(-z^2) * erfi(z)
}

