% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Voigt.R
\name{Gauss}
\alias{Gauss}
\title{Gauss}
\usage{
Gauss(x, x0, sigma)
}
\arguments{
\item{x}{numeric vector}

\item{x0}{scalar, peak position}

\item{sigma}{parameter of the gaussian}
}
\value{
numeric vector
}
\description{
Gaussian
}
\details{
Gaussian lineshape function
}
\author{
baptiste Auguie
}
\seealso{
Other helper_function: \code{\link{Lorentz}}
}

