% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Voigt.R
\name{Lorentz}
\alias{Lorentz}
\title{Lorentz}
\usage{
Lorentz(x, x0, gamma)
}
\arguments{
\item{x}{numeric vector}

\item{x0}{scalar, peak position}

\item{gamma}{parameter of the lorentzian}
}
\value{
numeric vector
}
\description{
Lorentzian
}
\details{
Lorentzian lineshape function
}
\author{
baptiste Auguie
}
\seealso{
Other helper_function: \code{\link{Gauss}}
}

