% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Voigt.R
\name{Voigt}
\alias{Voigt}
\title{Voigt}
\usage{
Voigt(x, x0, sigma, gamma, real = TRUE, ...)
}
\arguments{
\item{x}{numeric vector}

\item{x0}{scalar, peak position}

\item{sigma}{parameter of the gaussian}

\item{real}{logical, return only the real part of the complex Faddeeva}

\item{...}{passed to Faddeeva_w}

\item{sigma}{parameter of the lorentzian}
}
\value{
numeric or complex vector
}
\description{
Voigt distribution
}
\details{
The Voigt function, corresponding to the convolution of a lorentzian and a gaussian distribution
}
\author{
baptiste Auguie
}

