% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erf}
\alias{erf}
\title{compute erf(z), the error function of complex arguments}
\usage{
erf(z, relerr = 0)
}
\arguments{
\item{z}{complex vector}

\item{relerr}{double, requested error}
}
\value{
complex vector
}
\description{
compute erf(z), the error function of complex arguments
}

