% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erfc}
\alias{erfc}
\title{compute erfc(z) = 1 - erf(z), the complementary error function}
\usage{
erfc(z, relerr = 0)
}
\arguments{
\item{z}{complex vector}

\item{relerr}{double, requested error}
}
\value{
complex vector
}
\description{
compute erfc(z) = 1 - erf(z), the complementary error function
}

