% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erfcx}
\alias{erfcx}
\title{compute erfcx(z) = exp(z^2) erfc(z)}
\usage{
erfcx(z, relerr = 0)
}
\arguments{
\item{z}{complex vector}

\item{relerr}{double, requested error}
}
\value{
complex vector
}
\description{
compute erfcx(z) = exp(z^2) erfc(z)
}

