% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erfi}
\alias{erfi}
\title{compute erfi(z) = -i erf(iz), the imaginary error function}
\usage{
erfi(z_, relerr = 0)
}
\arguments{
\item{relerr}{double, requested error}

\item{z}{complex vector}
}
\value{
complex vector
}
\description{
compute erfi(z) = -i erf(iz), the imaginary error function
}

