% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Faddeeva_w}
\alias{Faddeeva_w}
\title{compute w(z) = exp(-z^2) erfc(-iz) (Faddeeva function)}
\usage{
Faddeeva_w(z, relerr = 0)
}
\arguments{
\item{z}{complex vector}

\item{relerr}{double, requested error}
}
\value{
complex vector
}
\description{
compute w(z) = exp(-z^2) erfc(-iz) (Faddeeva function)
}

