% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppFaddeeva-package.r
\docType{package}
\name{RcppFaddeeva-package}
\alias{RcppFaddeeva-package}
\title{RcppFaddeeva}
\description{
Rcpp Bindings for the "Faddeeva" Library
}
\details{
Access to error functions for arbitrary complex arguments is provided via the Faddeeva package by Steven G. Johnson.
}
\references{
The Faddeeva Package wiki page details the algorithms implemented by Steve G. Johnson,
\url{http://ab-initio.mit.edu/wiki/index.php/Faddeeva_Package}
}
\keyword{packagelibrary}

