% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Voigt.R
\name{Voigt}
\alias{Voigt}
\title{Voigt}
\usage{
Voigt(x, x0, sigma, gamma, real = TRUE, ...)
}
\arguments{
\item{x}{numeric vector}

\item{x0}{scalar, peak position}

\item{sigma}{parameter of the gaussian}

\item{gamma}{parameter of the lorentzian}

\item{real}{logical, return only the real part of the complex Faddeeva}

\item{...}{passed to Faddeeva_w}
}
\value{
numeric or complex vector
}
\description{
Voigt distribution
}
\details{
The Voigt function, corresponding to the convolution of a lorentzian and a gaussian distribution
}
\examples{
## should integrate to 1 in all cases
integrate(Voigt, -Inf, Inf, x0=400, sigma=50, gamma=100)
x <- seq(-1000, 1000)
x0 <- 200
l <- RcppFaddeeva:::Lorentz(x, x0, 30)
g <- RcppFaddeeva:::Gauss(x, x0, 100)
N <- length(x)
c <- convolve(Gauss(x, 200, 100),
              Lorentz(x, 0, 30), type="o")[seq(N/2, length=N)]
v <- Voigt(x, x0, 100, 30)
matplot(x, cbind(v, l, g, c), t="l", lty=c(1,2,2,1), xlab="x", ylab="")
legend("topleft", legend = c("Voigt", "Lorentz", "Gauss", "Convolution"), bty="n",
      lty=c(1,2,2,1), col=1:4)
}
\author{
baptiste Auguie
}

