% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{erfcx}
\alias{Dawson}
\alias{Faddeeva_w}
\alias{erf}
\alias{erfc}
\alias{erfcx}
\alias{erfi}
\title{compute erfcx(z) = exp(z^2) erfc(z)}
\usage{
erfcx(z, relerr = 0)

erf(z, relerr = 0)

erfi(z, relerr = 0)

erfc(z, relerr = 0)

Dawson(z, relerr = 0)

Faddeeva_w(z, relerr = 0)
}
\arguments{
\item{z}{complex vector}

\item{relerr}{double, requested error}

\item{z}{complex vector}

\item{relerr}{double, requested error}

\item{z}{complex vector}

\item{relerr}{double, requested error}

\item{z}{complex vector}

\item{relerr}{double, requested error}

\item{z}{complex vector}

\item{relerr}{double, requested error}

\item{z}{complex vector}

\item{relerr}{double, requested error}
}
\value{
complex vector

complex vector

complex vector

complex vector

complex vector

complex vector
}
\description{
compute erfcx(z) = exp(z^2) erfc(z)

compute erf(z), the error function of complex arguments

compute erfi(z) = -i erf(iz), the imaginary error function

compute erfc(z) = 1 - erf(z), the complementary error function

compute Dawson(z) = sqrt(pi)/2  *  exp(-z^2) * erfi(z)

compute w(z) = exp(-z^2) erfc(-iz) (Faddeeva function)
}
\section{Functions}{
\itemize{
\item \code{erfcx}: 

\item \code{erf}: 

\item \code{erfi}: 

\item \code{erfc}: 

\item \code{Dawson}: 

\item \code{Faddeeva_w}: 
}}
\examples{
erfcx(1:10 + 1i)
erf(1:10 + 1i)
erfi(1:10 + 1i)
erfc(1:10 + 1i)
Dawson(1:10 + 1i)
Faddeeva_w(1:10 + 1i)
}

