% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Faddeeva_w}
\alias{Faddeeva_w}
\alias{erfcx}
\alias{erf}
\alias{erfi}
\alias{erfc}
\alias{Dawson}
\title{Faddeeva family of error functions of the complex variable}
\usage{
Faddeeva_w(z, relerr = 0)

erfcx(z, relerr = 0)

erf(z, relerr = 0)

erfi(z, relerr = 0)

erfc(z, relerr = 0)

Dawson(z, relerr = 0)
}
\arguments{
\item{z}{complex vector}

\item{relerr}{double, requested error}
}
\value{
complex vector
}
\description{
the Faddeeva function
}
\section{Functions}{
\itemize{
\item \code{Faddeeva_w()}: compute \eqn{w(z) = \exp(-z^2) \, \mathrm{erfc}(-iz)}{w(z) = exp(-z^2) erfc(-iz)}

\item \code{erfcx()}: compute \eqn{\mathrm{erfcx}(z) = \exp(z^2) \, \mathrm{erfc}(z)}{erfcx(z) = exp(z^2) erfc(z)}

\item \code{erf()}: compute \eqn{\mathrm{erf}(z)}{erf(z)}

\item \code{erfi()}: compute \eqn{\mathrm{erfi}(z) = -i \, \mathrm{erf}(iz)}{erfi(z) = -i erf(iz)}

\item \code{erfc()}: compute \eqn{\mathrm{erfc}(z) = 1 - \mathrm{erf}(z)}{erfc(z) = 1 - erf(z)}

\item \code{Dawson()}: compute \eqn{\mathrm{Dawson}(z) = \sqrt{\pi}/2 \exp(-z^2) \, \mathrm{erfi}(z)}{Dawson(z) = sqrt(pi)/2 * exp(-z^2) * erfi(z)}

}}
\examples{
Faddeeva_w(1:10 + 1i)
erfcx(1:10 + 1i)
erf(1:10 + 1i)
erfi(1:10 + 1i)
erfc(1:10 + 1i)
Dawson(1:10 + 1i)
}
\concept{wrapper}
